python: moved sum_xxx, ecl_smspec, ecl_npv and ecl_cmp to ecl.summary.
tied all sum files to ecl.summary.

...
