+++ bin/summary_resample
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 import sys
 import argparse
-from ecl.ecl import EclSum
+from ecl.summary import EclSum
 from ecl.util import TimeVector, CTime
