+++ python/ecl/CMakeLists.txt
@@ -10,7 +10,7 @@ add_subdirectory(grid)
 add_subdirectory(faults)
 add_subdirectory(rft)
 add_subdirectory(gravimetry)
-#add_subdirectory(ecl)
+add_subdirectory(summary)
 add_subdirectory(geo)
 add_subdirectory(test)
 add_subdirectory(util)
