+++ python/ecl/summary/ecl_cmp.py
@@ -15,7 +15,7 @@
 #  for more details.
 
 from ecl.util import monkey_the_camel
-from ecl.ecl import EclSum
+from ecl.summary import EclSum
 
 class EclCase(object):
