+++ python/ecl/summary/ecl_smspec_node.py
