+++ python/ecl/summary/ecl_sum_keyword_vector.py
