+++ python/ecl/summary/ecl_sum_node.py
