+++ python/ecl/summary/ecl_sum_var_type.py
