+++ python/tests/ecl_tests/test_ecl_sum.py
@@ -19,8 +19,8 @@ import datetime
 import os.path
 
 from cwrap import CFILE
-from ecl.ecl import EclSum, EclSumKeyWordVector, EclFile
-from ecl.ecl import FortIO, openFortIO, openEclFile, EclKW
+from ecl.eclfile import EclFile, FortIO, openFortIO, openEclFile, EclKW
+from ecl.summary import EclSum, EclSumKeyWordVector
 from ecl.test import TestAreaContext
 from tests import EclTest, statoil_test
