+++ python/tests/ecl_tests/test_ecl_sum_vector.py
@@ -22,7 +22,7 @@ except ImportError:
 
 import warnings
 
-from ecl.ecl import EclSumVector, EclSum
+from ecl.summary import EclSumVector, EclSum
 from tests import EclTest, statoil_test
 
 @statoil_test()
